+++ lib/src/dependency_analysis.dart
@@ -1,9 +1,7 @@
 // lib/src/dependency_analysis.dart
 import 'dart:io';
 
-import 'package:analyzer/dart/analysis/utilities.dart';
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:path/path.dart' as path;
 
 import 'file_analyzer.dart';
 
@@ -13,11 +11,9 @@ class DependencyAnalysis implements FileAnalyzer {
 
   @override
   void analyzeFile(
-      File file, String content, CompilationUnit? compilationUnit) {
+      File file, String content, CompilationUnit? compilationUnit,
+      {required String relativePath}) {
     if (compilationUnit == null) return;
-
-    final relativePath = path.join('lib',
-        path.relative(file.path, from: path.dirname(path.dirname(file.path))));
 
     for (final directive in compilationUnit.directives) {
       if (directive is ImportDirective) {
@@ -28,15 +24,6 @@ class DependencyAnalysis implements FileAnalyzer {
         }
       }
     }
-  }
-
-  /// Analyzes dependencies for a single Dart file.
-  @Deprecated('Use analyzeFile instead')
-  void analyzeDependencies(File file) {
-    final content = file.readAsStringSync();
-    final result =
-        parseString(content: content);
-    analyzeFile(file, content, result.unit);
   }
 
   @override